// client/src/lib/shared/bindings/CardEntity.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { IncantationEntity } from "./IncantationEntity";
import type { MinionEntity } from "./MinionEntity";

export type CardEntity = { "Minion": MinionEntity } | { "Incantation": IncantationEntity };