// client/src/lib/shared/bindings/GameStateClient.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { Board } from "./Board";

export type GameStateClient = { self_board: Board, enemy_board: Board, white_player_id: string, black_player_id: string, your_turn: boolean, turn_count: number, cards_played_this_turn: number, };