// client/src/lib/shared/bindings/MinionEntity.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { MinionCard } from "./MinionCard";

export type MinionEntity = { entity_id: string, cost: number, turns_in_hand: number, just_drawn: boolean, card: MinionCard, attack: number, defence: number, exhausted: boolean, };